refactor: move broadcastMessageFromData to notifyListener
