+++ apps/meteor/app/discussion/server/hooks/propagateDiscussionMetadata.ts
@@ -2,15 +2,15 @@ import type { IRoom } from '@rocket.chat/core-typings';
 import { Messages, Rooms } from '@rocket.chat/models';
 
 import { callbacks } from '../../../../lib/callbacks';
-import { broadcastMessageFromData } from '../../../../server/modules/watchers/lib/messages';
 import { deleteRoom } from '../../../lib/server/functions/deleteRoom';
+import { notifyOnMessageChange } from '../../../lib/server/lib/notifyListener';
 
 const updateAndNotifyParentRoomWithParentMessage = async (room: IRoom): Promise<void> => {
 	const { value: parentMessage } = await Messages.refreshDiscussionMetadata(room);
 	if (!parentMessage) {
 		return;
 	}
-	void broadcastMessageFromData({
+	void notifyOnMessageChange({
 		id: parentMessage._id,
 		data: parentMessage,
 	});
