+++ apps/meteor/app/lib/server/functions/sendMessage.ts
@@ -7,11 +7,10 @@ import { Match, check } from 'meteor/check';
 import { callbacks } from '../../../../lib/callbacks';
 import { isRelativeURL } from '../../../../lib/utils/isRelativeURL';
 import { isURL } from '../../../../lib/utils/isURL';
-import { broadcastMessageFromData } from '../../../../server/modules/watchers/lib/messages';
 import { hasPermissionAsync } from '../../../authorization/server/functions/hasPermission';
 import { FileUpload } from '../../../file-upload/server';
 import { settings } from '../../../settings/server';
-import { notifyOnRoomChangedById } from '../lib/notifyListener';
+import { notifyOnRoomChangedById, notifyOnMessageChange } from '../lib/notifyListener';
 import { validateCustomMessageFields } from '../lib/validateCustomMessageFields';
 import { parseUrlsInMessage } from './parseUrlsInMessage';
 
@@ -292,7 +291,7 @@ export const sendMessage = async function (user: any, message: any, room: any, u
 
 	await callbacks.run('afterSaveMessage', message, room);
 
-	void broadcastMessageFromData({
+	void notifyOnMessageChange({
 		id: message._id,
 	});
