+++ apps/meteor/app/lib/server/functions/updateMessage.ts
@@ -5,9 +5,8 @@ import { Messages, Rooms } from '@rocket.chat/models';
 import { Meteor } from 'meteor/meteor';
 
 import { callbacks } from '../../../../lib/callbacks';
-import { broadcastMessageFromData } from '../../../../server/modules/watchers/lib/messages';
 import { settings } from '../../../settings/server';
-import { notifyOnRoomChangedById } from '../lib/notifyListener';
+import { notifyOnRoomChangedById, notifyOnMessageChange } from '../lib/notifyListener';
 import { validateCustomMessageFields } from '../lib/validateCustomMessageFields';
 import { parseUrlsInMessage } from './parseUrlsInMessage';
 
@@ -102,7 +101,7 @@ export const updateMessage = async function (
 		// so we wait for it to run before broadcasting
 		const data = await callbacks.run('afterSaveMessage', msg, room, user._id);
 
-		void broadcastMessageFromData({
+		void notifyOnMessageChange({
 			id: msg._id,
 			data: data as any, // TODO move "afterSaveMessage" type definition to specify a return value
 		});
