+++ apps/meteor/app/lib/server/lib/notifyListener.ts
@@ -17,6 +17,9 @@ import type {
 	AtLeast,
 	ISettingColor,
 	IUser,
+	IMessage,
+	SettingValue,
+	MessageTypesValues,
 } from '@rocket.chat/core-typings';
 import {
 	Rooms,
@@ -30,7 +33,11 @@ import {
 	LivechatInquiry,
 	LivechatDepartmentAgents,
 	Users,
+	Messages,
 } from '@rocket.chat/models';
+import mem from 'mem';
+
+import { shouldHideSystemMessage } from '../../../../server/lib/systemMessage/hideSystemMessage';
 
 type ClientAction = 'inserted' | 'updated' | 'removed';
 
@@ -401,3 +408,65 @@ export const notifyOnUserChangeById = withDbWatcherCheck(
 		void notifyOnUserChange({ id, clientAction, data: user });
 	},
 );
+
+const getUserNameCached = mem(
+	async (userId: string): Promise<string | undefined> => {
+		const user = await Users.findOne<Pick<IUser, 'name'>>(userId, { projection: { name: 1 } });
+		return user?.name;
+	},
+	{ maxAge: 10000 },
+);
+
+const getSettingCached = mem(async (setting: string): Promise<SettingValue> => Settings.getValueById(setting), { maxAge: 10000 });
+
+export async function getMessageToBroadcast({ id, data }: { id: IMessage['_id']; data?: IMessage }): Promise<IMessage | void> {
+	const message = data ?? (await Messages.findOneById(id));
+	if (!message) {
+		return;
+	}
+
+	if (message.t) {
+		const hiddenSystemMessages = (await getSettingCached('Hide_System_Messages')) as MessageTypesValues[];
+		const shouldHide = shouldHideSystemMessage(message.t, hiddenSystemMessages);
+
+		if (shouldHide) {
+			return;
+		}
+	}
+
+	if (message._hidden || message.imported != null) {
+		return;
+	}
+
+	const useRealName = (await getSettingCached('UI_Use_Real_Name')) === true;
+	if (useRealName) {
+		if (message.u?._id) {
+			const name = await getUserNameCached(message.u._id);
+			if (name) {
+				message.u.name = name;
+			}
+		}
+
+		if (message.mentions?.length) {
+			for await (const mention of message.mentions) {
+				const name = await getUserNameCached(mention._id);
+				if (name) {
+					mention.name = name;
+				}
+			}
+		}
+	}
+
+	return message;
+}
+
+export const notifyOnMessageChange = withDbWatcherCheck(async ({ id, data }: { id: IMessage['_id']; data?: IMessage }): Promise<void> => {
+	if (!dbWatchersDisabled) {
+		return;
+	}
+	const message = await getMessageToBroadcast({ id, data });
+	if (!message) {
+		return;
+	}
+	void api.broadcast('watch.messages', { message });
+});
