+++ apps/meteor/app/threads/server/hooks/aftersavemessage.ts
@@ -4,7 +4,7 @@ import { Messages } from '@rocket.chat/models';
 import { Meteor } from 'meteor/meteor';
 
 import { callbacks } from '../../../../lib/callbacks';
-import { broadcastMessageFromData } from '../../../../server/modules/watchers/lib/messages';
+import { notifyOnMessageChange } from '../../../lib/server/lib/notifyListener';
 import { updateThreadUsersSubscriptions, getMentions } from '../../../lib/server/lib/notifyUsersOnMessage';
 import { sendMessageNotifications } from '../../../lib/server/lib/sendNotificationsOnMessage';
 import { settings } from '../../../settings/server';
@@ -62,7 +62,7 @@ export async function processThreads(message: IMessage, room: IRoom) {
 	await notifyUsersOnReply(message, replies);
 	await metaData(message, parentMessage, replies);
 	await notification(message, room, replies);
-	void broadcastMessageFromData({
+	void notifyOnMessageChange({
 		id: message.tmid,
 	});
