+++ apps/meteor/server/modules/watchers/watchers.module.ts
@@ -45,16 +45,16 @@ type BroadcastCallback = <T extends keyof EventSignatures>(event: T, ...args: Pa
 
 const hasKeys =
 	(requiredKeys: string[]): ((data?: Record<string, any>) => boolean) =>
-	(data?: Record<string, any>): boolean => {
-		if (!data) {
-			return false;
-		}
+		(data?: Record<string, any>): boolean => {
+			if (!data) {
+				return false;
+			}
 
-		return Object.keys(data)
-			.filter((key) => key !== '_id')
-			.map((key) => key.split('.')[0])
-			.some((key) => requiredKeys.includes(key));
-	};
+			return Object.keys(data)
+				.filter((key) => key !== '_id')
+				.map((key) => key.split('.')[0])
+				.some((key) => requiredKeys.includes(key));
+		};
 
 const hasRoomFields = hasKeys(Object.keys(roomFields));
 const hasSubscriptionFields = hasKeys(Object.keys(subscriptionFields));
